requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
chromadb>=0.5.0
sentence-transformers>=2.6.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pypdf>=4.2.0
